🧪 KANFC0 code: isolate Windows identity regression
